+++ pkg/interpreter/interpreter_integration_test.go
@@ -286,6 +286,35 @@ var integrationTests = []struct {
 		wantCategory:   "",
 		config:         interpreter.DefaultConfig(),
 	},
+	// v0.26.0 regression tests
+	{
+		name:           "time ticker",
+		dir:            "time_ticker",
+		wantViolations: 0,
+		wantCategory:   "",
+		config:         interpreter.DefaultConfig(),
+	},
+	{
+		name:           "os file rw",
+		dir:            "os_file_rw",
+		wantViolations: 0,
+		wantCategory:   "",
+		config:         interpreter.DefaultConfig(),
+	},
+	{
+		name:           "http client",
+		dir:            "http_client",
+		wantViolations: 0,
+		wantCategory:   "",
+		config:         interpreter.DefaultConfig(),
+	},
+	{
+		name:           "signal notify",
+		dir:            "signal_notify",
+		wantViolations: 0,
+		wantCategory:   "",
+		config:         interpreter.DefaultConfig(),
+	},
 	// v0.25.0 regression tests
 	{
 		name:           "url parse",